requirements.txt
fastapi
uvicorn
websockets
opencv-python
numpy
pillow
pyautogui
pydirectinput
flask
torch
torchvision
tbb
